requirements.txt
Django==4.1.4
django_birthday==0.1.4
django_dob_widget==1.2.0
djangorestframework==3.14.0
sweetify==2.3.1
xlwt==1.3.0
